requirements.txt
# ── Core ──────────────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# ── Deep Learning ─────────────────────────────────────────────
torch>=2.1.0
pytorch-lightning>=2.1.0
torchmetrics>=1.2.0

# ── Hyperparameter Tuning ─────────────────────────────────────
optuna>=3.4.0
optuna-integration[pytorch-lightning]

# ── Data Collection ───────────────────────────────────────────
ccxt>=4.1.0
requests>=2.31.0
yfinance>=0.2.31
pyarrow>=14.0.0

# ── Feature Engineering ───────────────────────────────────────
ta>=0.11.0
PyWavelets>=1.4.1
statsmodels>=0.14.0

# ── NLP / Sentiment ───────────────────────────────────────────
transformers>=4.35.0
torch-sentiment>=0.1.0
vaderSentiment>=3.3.2

# ── Experiment Tracking ───────────────────────────────────────
mlflow>=2.8.0
wandb>=0.16.0

# ── Visualization ─────────────────────────────────────────────
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
shap>=0.44.0

# ── Utilities ─────────────────────────────────────────────────
python-dotenv>=1.0.0
pyyaml>=6.0.1
loguru>=0.7.2
rich>=13.7.0
tqdm>=4.66.0
joblib>=1.3.2

# ── Testing ───────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0
